undo moving characters etc. to /search; add search bar to landing page banner
