+++ frontend/src/components/pages/index/LandingPageBanner.js
@@ -1,10 +1,12 @@
 import Image from "next/image";
 
 import { Box, Card, CardContent, Typography } from "@mui/material";
+import SearchBar from "./SearchBar";
 
 const LandingPageBanner = () => {
   return (
     <Card
+      id="db-search"
       sx={{
         position: "relative",
         borderRadius: "10px",
@@ -46,6 +48,7 @@ const LandingPageBanner = () => {
               An API with Data from the Harry Potter Universe.
             </Typography>
           </Box>
+          <SearchBar style={{ mt: 4 }} />
         </CardContent>
       </Box>
     </Card>
